consolidate sft dataloader tests
